refactor: delete old api calls
