+++ src/services/apiSlice.ts
@@ -1,5 +1,48 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 
+interface Character {
+  uid: string;
+  name: string;
+  gender: string | null;
+  yearOfBirth: number | null;
+  monthOfBirth: number | null;
+  dayOfBirth: number | null;
+  placeOfBirth: string | null;
+  yearOfDeath: number | null;
+  monthOfDeath: number | null;
+  dayOfDeath: number | null;
+  placeOfDeath: string | null;
+  height: number | null;
+  weight: number | null;
+  deceased: true | null;
+  bloodType: string | null;
+  maritalStatus: string | null;
+  serialNumber: string | null;
+  hologramActivationDate: string | null;
+  hologramStatus: string | null;
+  hologramDateStatus: string | null;
+  hologram: boolean | null;
+  fictionalCharacter: boolean | null;
+  mirror: boolean | null;
+  alternateReality: boolean | null;
+}
+
+interface StartrekApiResponse {
+  page: {
+    pageNumber: number;
+    pageSize: number;
+    numberOfElements: number;
+    totalElements: number;
+    totalPages: number;
+    firstPage: boolean;
+    lastPage: boolean;
+  };
+  sort: {
+    clauses: [[]] | never[];
+  };
+  characters: Character[] | never[];
+}
+
 const baseURL = "https://stapi.co/api/v1/rest/character";
 
 export const apiSlice = createApi({
@@ -32,3 +75,4 @@ export const apiSlice = createApi({
 
 export const { useGetCharactersQuery, useGetDetailedCharacterDataQuery } =
   apiSlice;
+export type { Character, StartrekApiResponse };
